requirements.txt
pandas
numpy
scikit-learn
CatBoostClassifier
psycopg2
apache-airflow
